% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RANDmillion-package.R
\docType{package}
\name{RANDmillion-package}
\alias{RANDmillion}
\alias{RANDmillion-package}
\title{RANDmillion: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Pete Dodd \email{p.j.dodd@sheffield.ac.uk}

}
\keyword{internal}
